Created question tags
The names of each tag are stored lowercase and separated by commas in the tag_names field of each question model. There is also a tags field in the question tag containing the ids for the QuestionTag model. When adding tags to a question during a POST or PUT request, do NOT edit the tags field as it is automatically cleared and filled with the corresponding ids from the names declared in tag_names.
Therefore, if you want to edit the tags in a question model, simply edit the string in tag_names, separating each tag name with a comma.