Completed all model changes
Also includes changes to tests and views.
Pictures are their own separate models as this is the best way I could find to save pictures of different types at different locations and for the client not to have to upload an image every time they wanted to update a profile, question or solution.