Implemented changes to backend
All params to do with UserProfile or User model instances require the token key as the value

Rank calculation happens on a separate thread and is an algorithm that runs every 10 seconds. When the main thread is stopped by keyboard interrupt, make sure to wait a few seconds for it will take the same interval for this thread to finish.

Number of questions to be chosen randomly is passed through request params with the key 'number'

Added QuestionFilterResult model

All models that are not UserProfile or User are linked to a UserProfile when needed with a foreign key field of name user_profile

Ensured password arguments for the creation of new User instaces are encrypted